+++ packages/yourloops/components/dashboard-cards/devices/device-list-card.tsx
@@ -30,9 +30,12 @@ import Typography from '@mui/material/Typography'
 import Box from '@mui/material/Box'
 import { DataCard } from '../../data-card/data-card'
 import { useTranslation } from 'react-i18next'
-import { PumpSettings } from 'medical-domain'
+import { PumpSettings, Unit } from 'medical-domain'
 import { makeStyles } from 'tss-react/mui'
 import { useTheme } from '@mui/material/styles'
+import { DownloadDocumentButton } from '../../buttons/download-document-button'
+import { usePatient } from '../../../lib/patient/patient.provider'
+import { getIfuDocumentName } from '../../../lib/medical-files/medical-files.utils'
 
 interface DeviceListCardProps {
   pumpSettings: PumpSettings
@@ -45,10 +48,13 @@ const useStyles = makeStyles()(() => ({
 }))
 
 export const DeviceListCard: FC<DeviceListCardProps> = (props) => {
-  const { t } = useTranslation()
+  const { t, i18n } = useTranslation()
   const { pumpSettings } = props
   const { classes } = useStyles()
   const theme = useTheme()
+  const { patient } = usePatient()
+  const bgUnit = patient?.settings?.units?.bg ?? Unit.MilligramPerDeciliter
+  const ifuDocumentName = pumpSettings ? getIfuDocumentName(pumpSettings, bgUnit, i18n.language) : null
 
   const deviceData = {
     cgm: {
@@ -57,19 +63,32 @@ export const DeviceListCard: FC<DeviceListCardProps> = (props) => {
     },
     device: {
       label: t('dbl'),
-      value: pumpSettings?.payload.device.manufacturer ?? ''
+      value: pumpSettings?.payload.device.name ?? ''
     },
     pump: {
       label: t('Pump'),
-      value: pumpSettings?.payload.pump.manufacturer ?? ''
+      value: pumpSettings?.payload.pump.manufacturer ? `${pumpSettings.payload.pump.manufacturer} ${pumpSettings.payload.pump.name}` : ''
     }
   }
 
   return (
     <DataCard data-testid="device-usage-device-list">
-      <Typography sx={{ fontWeight: 'bold', paddingBottom: theme.spacing(1) }}>
-        {t('devices')}
-      </Typography>
+      <Box sx={{ paddingBottom: theme.spacing(1), display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
+        <Typography sx={{ fontWeight: 'bold' }}>
+          {t('devices')}
+        </Typography>
+        {ifuDocumentName &&
+          <DownloadDocumentButton
+            documentName={ifuDocumentName}
+            metricName="dashboard-download-ifu"
+            labelKey="button-download-ifu"
+            size="medium"
+          />
+        }
+
+      </Box>
+
+
       {Object.keys(deviceData).map(
         (key) =>
           <React.Fragment key={key}>