+++ packages/yourloops/test/integration/pages/patient-view/dashboard/dashboard-view.hcp.spec.tsx
@@ -192,7 +192,7 @@ describe('Dashboard view for HCP', () => {
       renderPage(patientDashboardRoute)
     })
 
-    await testDashboardDataVisualisationTwoWeeksOldData()
+    await testDashboardDataVisualisationTwoWeeksOldData(false)
   })
 
   it('should render correct components when patient is in no medical teams', async () => {
@@ -227,7 +227,7 @@ describe('Dashboard view for HCP', () => {
 
     await testAppMainLayoutForHcp(appMainLayoutParams)
     testPatientNavBarLayoutForHcpInPrivateTeam()
-    await testDashboardDataVisualisationPrivateTeamNoData(patientDashboardLayoutParams)
+    await testDashboardDataVisualisationPrivateTeamNoData(patientDashboardLayoutParams, false)
   })
 
   it('should automatically log out an idle user', async () => {