+++ CHANGELOG.md
@@ -7,6 +7,7 @@ It is based on Tidepool Blip 1.27.
 - YLP-3853 Display target value on Daily Glucose chart
 - YLP-3892 Accept data sharing request from external partners
 - YLP-3893 Data sharing management with external partners
+- YLP-3904 As an HCP or a patient, I would like to be able to download the IFU corresponding to the system used, directly from Yourloops
 - YLP-3907 Add a conditional rendering on birth date display
 
 ### Changed